+++ packages/js-client-core/src/telemetry/types.ts
@@ -9,6 +9,8 @@ export type EventReport = {
   metaContext: {
     sdkName: string;
     sdkVersion: string;
+    appName?: string;
+    appVersion?: string;
   },
   discreteEvents: DiscreteEventList;
   aggregatedEvents: AggregatedEventList;