+++ packages/nuxt-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @configdirector/nuxt-sdk
 
+## 1.6.1
+
+### Patch Changes
+
+- Fixed telemetry reporter to include appName and appVersion with telemetry data to support activity graphs.
+
 ## 1.6.0
 
 ### Minor Changes