+++ packages/openfeature-server-provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @configdirector/openfeature-server-provider
 
+## 1.6.1
+
+### Patch Changes
+
+- Fixed telemetry reporter to include appName and appVersion with telemetry data to support activity graphs.
+
 ## 1.6.0
 
 ### Minor Changes